pyproject.toml
[tool.poetry]
name = "kafka-clickhouse-project"
version = "0.1.0"
description = "A data pipeline project using Kafka and ClickHouse"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
kafka-python = "^2.0.2"
clickhouse-driver = "^0.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
mypy = "^0.950"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 